Fitxa de regles refeta, so sintetitzat i amplada dels fulls
Les regles es pintaven gairebé sense estil: .regles-seccio, .ordre,
.valors i .btn eren classes de la versió anterior del full d'estil i ja no
existien. Refetes amb el sistema actual: jerarquia com a tira de cartes
reals desplaçable, taula de valors amb cromos, les cinc obligacions
numerades i els multiplicadors. Fan servir la baralla que tinguis triada.

De passada, un error que afectava tots els fulls flotants: .capa-modal
.full posava left:50% però la regla base deixava right:0, i l'amplada
quedava a la meitat del contenidor. Els ajustos i el menú de partida
també en sortien estrets.

So: tot sintetitzat amb Web Audio, sense cap fitxer ni cap llicència.
Clic de botó, carta sobre el tapet, repartiment, recollida de basa, cant,
contro i acord de final de mà, amb el volum al 40% i ±4% de to aleatori
per no sonar a màquina, tal com diu la §11.

I una fanfàrria de victòria amb redoblament i metalls. Va contra la §11 i
la §15, que demanen sobrietat; hi és per decisió expressa, i sona només en
guanyar la partida, no cada mà.

+++ server/src/api.rs
@@ -288,7 +288,8 @@ async fn estadistiques(estat: &Estat, usuari: Uuid) -> Resultat<Value> {
            coalesce(sum(mans_guanyades), 0),
            coalesce(sum(nos), 0),
            coalesce(sum(ells), 0),
-           coalesce(max(millor_ma), 0)
+           -- max() d'un int4 torna int4, no int8: cal dir-ho.
+           coalesce(max(millor_ma), 0)::bigint
          from partides where usuari = $1",
     )
     .bind(usuari)
